Can now pull data from the database
Pulls data based on UUID, along with associated worlds and their stats. This can be seen in the test.js page, though it has no button to access it. To view it I would suggest temporarily changing one of the buttons on the index page to route to test. Right now it pulls pre-seeded data from the database though I'm sure it would work once we get the local app running.